Remove deprecated email capture feature (#562)
* Remove email capture feature

* Fix tests

* Remove unused style

* Remove more text keys from reducer
