+++ dev-site/embedded.html
@@ -20,7 +20,6 @@
       jwt: 'JWT',
       userId: 'USER_ID',
       properties: PROPERTIES,
-      emailCaptureEnabled: true,
       embedded: true
     };
