+++ dev-site/index.html
@@ -18,8 +18,7 @@
             email: 'EMAIL',
             jwt: 'JWT',
             userId: 'USER_ID',
-            properties: PROPERTIES,
-            emailCaptureEnabled: true
+            properties: PROPERTIES
         };
 
         Smooch.on('message:received', function(message) {
