+++ release_notes/v4.0.0-wip.md
@@ -1,0 +1,4 @@
+# What's new?
+
+- The Web Messenger is now run and rendered inside of an iframe in order to maximize compatibility with sites where it is used.
+- Email capture capability has been deprecated a while ago in favor of channel linking and was removed from the Web Messenger.
