+++ src/frame/js/components/settings.jsx
@@ -2,28 +2,66 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
-import { EmailSettings } from './email-settings';
-import { NotificationsSettings } from './notifications-settings';
-import { hasChannels } from '../utils/app';
+import { NotificationChannelItem } from './notification-channel-item';
+
+import { getAppChannelDetails } from '../utils/app';
+import { isChannelLinked, getDisplayName } from '../utils/user';
+
 
 export class SettingsComponent extends Component {
 
     static propTypes = {
-        settings: PropTypes.object.isRequired
+        appChannels: PropTypes.array.isRequired,
+        user: PropTypes.object.isRequired,
+        notificationSettingsChannelsTitleText: PropTypes.string.isRequired,
+        notificationSettingsChannelsDescriptionText: PropTypes.string.isRequired
     };
 
     render() {
-        const {settings} = this.props;
-        const settingsComponent = hasChannels(settings) ? <NotificationsSettings /> : <EmailSettings />;
+        const {appChannels, user, notificationSettingsChannelsTitleText, notificationSettingsChannelsDescriptionText} = this.props;
+
+        if (!user._id) {
+            return null;
+        }
+
+        let channels = getAppChannelDetails(appChannels);
+        channels.sort(({channel}) => {
+            // List the linked channels first
+            return isChannelLinked(user.clients, channel.type) ? -1 : 1;
+        });
+
+        channels = channels.map(({channel, details}) => {
+            return <NotificationChannelItem key={ channel.type }
+                                            id={ channel.type }
+                                            {...details}
+                                            displayName={ getDisplayName(user.clients, channel.type) }
+                                            linked={ isChannelLinked(user.clients, channel.type) }
+                                            hasURL={ !!details.getURL(channel) } />;
+        });
 
         return <div className='sk-settings'>
-                   { settingsComponent }
+                   <div className='content-wrapper'>
+                       <div className='settings-wrapper'>
+                           <p className='settings-header'>
+                               { notificationSettingsChannelsTitleText }
+                           </p>
+                           <p className='settings-description'>
+                               { notificationSettingsChannelsDescriptionText }
+                           </p>
+                           <div className='channels'>
+                               { channels }
+                           </div>
+                       </div>
+                   </div>
                </div>;
     }
 }
 
-export const Settings = connect(({app}) => {
+export const Settings = connect(({app, user, ui}) => {
     return {
-        settings: app.settings.web
+        appChannels: app.integrations,
+        notificationSettingsChannelsTitleText: ui.text.notificationSettingsChannelsTitle,
+        notificationSettingsChannelsDescriptionText: ui.text.notificationSettingsChannelsDescription,
+        user
     };
 })(SettingsComponent);
