+++ src/frame/js/reducers/ui-reducer.js
@@ -8,12 +8,7 @@ const INITIAL_STATE = {
         sendButtonText: 'Send',
         introductionText: 'We\'re here to talk, so ask us anything!',
         introAppText: 'Message us below or from your favorite app.',
-        settingsText: 'You can leave us your email so that we can get back to you this way.',
-        settingsReadOnlyText: 'We\'ll get back to you at this email address if we missed you.',
-        settingsInputPlaceholder: 'Your email address',
-        settingsSaveButtonText: 'Save',
         settingsHeaderText: 'Settings',
-        settingsNotificationText: 'In case we\'re slow to respond you can <a href data-ui-settings-link>leave us your email</a>.',
         actionPaymentError: 'An error occurred while processing the card. <br> Please try again or use a different card.',
         actionPaymentCompleted: 'Payment Completed',
         actionPostbackError: 'An error occurred while processing your action. Please try again.',
