+++ src/frame/stylesheets/settings.less
@@ -19,37 +19,6 @@
 
         box-sizing: border-box;
 
-        .input-group {
-            position: relative;
-
-            i.before-icon {
-                color: #bdbdbd;
-                position: absolute;
-                top: 16px;
-                left: 11px;
-            }
-
-            .email-input {
-                box-sizing: border-box;
-                padding: 0 @input-horizontal-padding 0 30px;
-                width: 100%;
-            }
-
-            .form-message {
-                .button-size(@padding-base-vertical; @padding-base-horizontal; @font-size-base; @line-height-base; @btn-border-radius-base);
-                vertical-align: middle;
-
-                i.success {
-                    color: @brand-success;
-                }
-            }
-
-            .btn {
-                width: 150px;
-                border-radius: 20px;
-            }
-        }
-
         .settings-header {
             font-weight: 700;
             font-size: 15px;
